Deal with Sublime Text 2 not properly returning when doing subl -nw on a file.
This issue occured most often when doing commits, since the editor for git
is "subl -nw". While running in standalone iTerm, it worked fine. When running
in tmux, after saving the file and closing Sublime, it acted like it never got
input back. Apparently it has to do with some pasteboard whackiness.